Remove whole tree, minus minus and note bug
We create a new tmpdir per file. Might be a bit inefficient but
avoids worrying about clashes. Although cleanup is set, I also
explicitely remove. No point in keeping the root.

Testing this using the source repo is hard. However I think it's
a real bug - if a file doesn't exist on the branch you have checked
out, any violations will be skipped because of the -e - needs
to change to a git show?